feat(display): introduce unified C_Display abstraction with OLED/LCD support and optimise Processor pipeline
- Added new C_OLED12864_SSD1306 component
- Implemented unified C_Display component with I_Display interface
- Added compile-time USE_OLED_DISPLAY flag for seamless LCD/OLED switching
- Updated C_Processor to use I_Display exclusively (no display branching)
- Updated C_Build to wire C_Display instead of individual display components
- Integrated SSD1306 OLED12864 support via Adafruit_GFX/SSD1306
- Added LCD1602 support via LiquidCrystal_I2C
- Refactored DHT22 adapter to return integer-tenths instead of floats
- Removed dtostrf(), snprintf(), and float formatting from Processor
- Added lightweight integer-based formatting functions
- Reduced stack usage and improved RAM stability on ATmega328P
- Cleaned up interfaces, adapters, and component wiring
- Verified correct rendering on both LCD and OLED targets
- Ensured full compatibility with existing architecture

+++ Arduino/RackFanController/src/C_DHT22.h
@@ -13,8 +13,8 @@ class I_DHT22 {
 public:
     virtual ~I_DHT22() = default;
     virtual void begin() = 0;
-    virtual float readTemperature() = 0;
-    virtual float readHumidity() = 0;
+    virtual int readTemperature() = 0;
+    virtual int readHumidity() = 0;
 };
 
 
@@ -27,8 +27,8 @@ public:
     A_DHT22(uint8_t pin, I_Arduino* p_IArduino);
 
     void begin() override;
-    float readTemperature() override;
-    float readHumidity() override;
+    int readTemperature() override;
+    int readHumidity() override;
 
 private:
     DHT dht;
